Update README and Example code
fixes #12
