+++ README.md
@@ -9,11 +9,14 @@ JSON Reference Implementation for Go
 # SYNOPSIS
 
 ```go
+package jsref_test
+
 import (
   "encoding/json"
+  "fmt"
   "log"
 
-  "github.com/lestrrat/go-jsref"
+  jsref "github.com/lestrrat/go-jsref"
   "github.com/lestrrat/go-jsref/provider"
 )
 
@@ -21,7 +24,7 @@ func Example() {
   var v interface{}
   src := []byte(`
 {
-  "foo": ["bar", {"$ref": "#/sub"}, {"$ref", "obj2#/sub"}],
+  "foo": ["bar", {"$ref": "#/sub"}, {"$ref": "obj2#/sub"}],
   "sub": "baz"
 }`)
   if err := json.Unmarshal(src, &v); err != nil {
@@ -36,9 +39,27 @@ func Example() {
   res := jsref.New()
   res.AddProvider(mp) // Register the provider
 
-  res.Resolve(v, "#/foo/0") // "bar"
-  res.Resolve(v, "#/foo/1") // "baz"
-  res.Resolve(v, "#/foo/2") // "quux" (resolve via `mp`)
+  ptrs := []string{
+    "#/foo/0", // "bar"
+    "#/foo/1", // "baz"
+    "#/foo/2", // "quux" (resolves via `mp`)
+    "#/foo",   // contents of foo key
+  }
+  for _, ptr := range ptrs {
+    result, err := res.Resolve(v, ptr)
+    if err != nil { // failed to resolve
+      fmt.Printf("err: %s\n", err)
+      continue
+    }
+    b, _ := json.Marshal(result)
+    fmt.Printf("%s -> %s\n", ptr, string(b))
+  }
+
+  // OUTPUT:
+  // #/foo/0 -> "bar"
+  // #/foo/1 -> "baz"
+  // #/foo/2 -> "quux"
+  // #/foo -> ["bar",{"$ref":"#/sub"},{"$ref":"obj2#/sub"}]
 }
 ```
